Also fixed memory print for background render to survive > 2 gig (unsigned)

+++ source/blender/render/intern/source/pipeline.c
@@ -124,7 +124,7 @@ static void print_error(const char *str) {printf("ERROR: %s\n", str);}
 
 static void stats_background(RenderStats *rs)
 {
-	extern int mem_in_use;
+	extern unsigned long mem_in_use;
 	float megs_used_memory= mem_in_use/(1024.0*1024.0);
 	char str[400], *spos= str;
